feat: add cancellation request handling and email notification
- Implemented logic to prevent deletion of important job positions in job-positions.service.ts.
- Updated seed data for job positions to reflect changes in roles and descriptions.
- Added a new email template for cancellation requests to approvers.
- Introduced a method in mail-client.service.ts to send cancellation request emails.
- Modified request approval entity and DTOs to remove unnecessary fields.
- Enhanced request approval repository to find current requests without cancellations.
- Refactored request approvals service to utilize new repository methods.
- Added cancellation functionality in requests service, allowing users to cancel requests with a reason.
- Updated request entity to include a field for cancellation reasons.
- Created a new DTO for cancellation requests.
- Adjusted request repository to support filtering requests by various criteria.
- Updated request controller to handle cancellation requests.
- Cleaned up commented-out code and improved overall code readability.

+++ src/job-positions/job-positions.service.ts
@@ -80,6 +80,17 @@ export class JobPositionsService {
         );
       }
 
+      // Check if the job position is ALCALDE or VICEALCALDE or ADMIN DE RECURSO HUMANO
+      const isImportantJobPosition =
+        jobPositionToRemove.id === 1 ||
+        jobPositionToRemove.id === 2 ||
+        jobPositionToRemove.id === 3;
+
+      if (isImportantJobPosition) {
+        throw new ConflictException(
+          'No se puede eliminar un este puesto de trabajo porque es necesario para el funcionamiento del sistema',
+        );
+      }
       return await this.jobPositionRepository.remove(jobPositionToRemove);
     } catch (error) {
       throw error;
